pyproject.toml
[tool.poetry]
name = "agro_app"
version = "0.1.0"
description = "Agro app"
authors = ["username <your@email.com>"]
packages = [
    {include = "**/*.py"},
]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.1"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
gevent = "^21.12.0"

## Extras
# test
ward = { version =  "^0.66.0-beta.0", optional = true }
coverage = {extras = ["toml"], version = "^6.3.2", optional = true }
ipdb = { version = "^0.13.9", optional = true }

# dev
pre-commit = { version = "^2.18.1", optional = true }
gitlint = { version = "^0.17.0", optional = true }

# lint
blue = { version = "^0.8.0", optional = true }
isort = { version = "^5.10.1", optional = true }
flake8 = { version = "^4.0.1", optional = true }
xenon = { version = "^0.9.0", optional = true }
mypy = { version = "^0.942", optional = true }

[tool.poetry.dev-dependencies]

[tool.poetry.extras]
test = [
    "ward",
    "ipdb",
]

dev = [
    "pre-commit",
    "gitlint",
]

lint = [
    "blue",
    "isort",
    "flake8",
    "xenon",
    "mypy",
]

[tool.isort]
multi_line_output = 3
line_length = 79
include_trailing_comma = true

[tool.coverage.run]
omit = [
  "agro_app/config.py"
]


[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = "agro_app/config.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
